[addition] skybox support for TextureCubePBR textures
TextureCubePBR textures provide multiple images w/ varying levels of
blur for roughness calculations, but they can still be passed in as
skybox textures since they are cube maps.

This commit adds detecting that to the skybox shader and pulling the
original unblurred portion when sampling to allow re-using these textures
instead of having to create a separate TextureCube for the skybox.