project folder restructure (see commit notes)
**Notice**

This project currently does NOT work with the newest GameMaker version
due to many breaking changes. This will be addressed in further commits.

**Restructure**

In preparation for C++ library support the project structure is being
adjusted so that GameMaker's project files are no longer at the root.
All source code will now be in the `./source/` where source files will
then be divided by type.