[addition] SSAO render size scaling
SSAO scaling is now added so that SSAO can be rendered at a separate
resolution from the primary buffers. The default scale is now set to 0.5
which results in equally good SSAO while having a much lower cost.

The scale can be lowered all the way down to 10% of the camera buffer
resolution.